# doc/api.yaml
openapi: 3.0.4
info:
  title: WASAText API
  description: Professional API for the functionalities of the WASAText messaging app
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            description: Error response for not found resources
            properties:
              error:
                type: string
                description: Error message
                example: Not found
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            type: object
            description: Error response for unauthorized access
            properties:
              error:
                type: string
                description: Error message
                example: Unauthorized
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            type: object
            description: Error response for invalid input data
            properties:
              error:
                type: string
                description: Error message
                example: Invalid input
  schemas:
    User:
      type: object
      description: A user of the WASAText app
      required: [id, username]
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username for the user
          example: johndoe
        displayName:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*$'
          description: Display name of the user
          example: John Doe
        profilePicture:
          type: string
          format: uri
          minLength: 10
          maxLength: 2048
          pattern: '^https?://.*$'
          description: URL of the user's profile picture
          example: https://example.com/avatar.jpg
...
paths:
  /session:
    post:
      summary: Logs in the user
      description: Allows a user to log in by providing their name.
      operationId: doLogin
      tags: [login]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: Request body for user login
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_]+$'
                  description: Name of the user
                  example: Maria
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
  /users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
    patch:
      summary: Update username
      description: Update the username of a user
      operationId: setMyUserName
      tags: [user]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for updating username
        content:
          application/json:
            schema:
              type: object
              required: [newName]
              properties:
                newName:
                  type: string
                  minLength: 3
                  maxLength: 32
                  pattern: '^[a-zA-Z0-9_]+$'
                  description: New username for the user
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    get:
      summary: Get a user by userId
      description: Retrieve a user by their userId
      operationId: getUser
      tags: [user]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userId}/photo:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
    patch:
      summary: Update user profile picture
      description: Update the profile picture of a user
      operationId: setMyPhoto
      tags: [user]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for updating user photo
        content:
          application/json:
            schema:
              type: object
              required: [photoUrl]
              properties:
                photoUrl:
                  type: string
                  format: uri
                  minLength: 10
                  maxLength: 2048
                  pattern: '^https?://.*$'
                  description: New photo URL for the user
      responses:
        '200':
          description: Profile picture updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /search/users:
    get:
      summary: Search users by username
      description: Search for users whose username contains the given query string.
      operationId: searchUsers
      tags: [user]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 32
            pattern: '^[a-zA-Z0-9_]+$'
          description: Partial or full username to search for
      responses:
        '200':
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                description: List of matching users
                minItems: 0
                maxItems: 1000
                items:
                  type: object
                  description: User found by search
                  properties:
                    id:
                      type: integer
                      description: Unique user identifier
                    username:
                      type: string
                      description: Username of the user
                    profilePicture:
                      type: string
                      description: URL of the user's profile picture
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...